ref(cartographer): remove `get_mode`
This function looked for each valid {mode} alphabetically within the
table. WHen it found one, it would return it. If none were found, it
would return an empty string.

Now, when indexing the table, if any key is one letter (e.g. `n` or `i`)
it is marked as the current mode. This mode is retrieved when assigning
a mapping, rather than iterating over all possible values.

This is a more future-proof implementation.